# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "EasyKos API Documentation",
        "description": "Dokumentasi API untuk aplikasi EasyKos - Manajemen Kos-kosan",
        "contact": {
            "email": "admin@easykos.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "EasyKos API Server Local"
        }
    ],
    "paths": {
...
        "/api/pemesanan": {
            "get": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Daftar pemesanan milik user yang login",
                "operationId": "6b77cf85c3bbd7ef77863d0565dd7d9a",
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil daftar pemesanan"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Buat pemesanan baru",
                "operationId": "e603a0cb60669e4a33f3389a873d9a2d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "kosan_id",
                                    "tanggal_masuk",
                                    "durasi_bulan"
                                ],
                                "properties": {
                                    "kosan_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "tanggal_masuk": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-01"
                                    },
                                    "durasi_bulan": {
                                        "type": "integer",
                                        "example": 3,
                                        "maximum": 24,
                                        "minimum": 1
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Mohon segera diproses."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pemesanan berhasil dibuat"
                    },
                    "400": {
                        "description": "Kamar penuh atau sudah ada pemesanan aktif"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/pemesanan/{id}": {
            "get": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Detail satu pemesanan",
                "operationId": "dadeddc1d36b5beae638cd73e48a669b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Berhasil mengambil detail pemesanan"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Pemesanan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Hapus pemesanan (hanya jika pending dan milik sendiri)",
                "operationId": "45170e1650d989b852447a6ee888921c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Pemesanan berhasil dihapus"
                    },
                    "400": {
                        "description": "Hanya pending yang bisa dihapus"
                    },
                    "404": {
                        "description": "Pemesanan tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Pemesanan"
                ],
                "summary": "Update status pemesanan",
                "description": "Pemilik bisa setuju/tolak, User bisa batalkan",
                "operationId": "921cd4ff6d0add81279bb4a14d2ec961",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID Pemesanan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "disetujui",
                                            "ditolak",
                                            "dibatalkan"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status berhasil diperbarui"
                    },
                    "400": {
                        "description": "Status tidak dapat diproses"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Masukkan token JWT Anda di sini. Cukup masukkan tokennya saja (tanpa kata 'Bearer ')",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}